+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Samsa Changelog
 
+## [Unreleased]
+### Changed
+- [#25](https://github.com/CallistoLabsNYC/samsa/issues/25) Offer a way to read from a Consumer without streaming
+
 ## [0.1.1] - 2024-03-02
 ### Fixed
 - [#20](https://github.com/CallistoLabsNYC/samsa/issues/20) Cannot perform ConsumerBuilder::build() via a reference