+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Fixed
 - [#20](https://github.com/CallistoLabsNYC/samsa/issues/20) Cannot perform ConsumerBuilder::build() via a reference
 - [#10](https://github.com/CallistoLabsNYC/samsa/issues/10) Export the TopicPartitions type
+- [#23](https://github.com/CallistoLabsNYC/samsa/issues/23) Update producer message version
 
 ## [0.1.0] - 2024-02-26
 ### Added